fix(test): fix iterator in awards test
